+++ src/EscapeWork/LaravelHelpers/Eloquent.php
@@ -2,30 +2,51 @@
 namespace EscapeWork\LaravelHelpers;
 
 use Illuminate\Database\Eloquent\Model as Model;
+use Cocur\Slugify\Slugify;
 use \Validator;
 use \Event;
 
 class Eloquent extends Model
 {
 
-    # regras de validações
+    /**
+     * Validation rules
+     * @var array
+     */
     public static $validationRules = array();
 
-    # mensagens das validações
+    /**
+     * Validation messages
+     * @var array
+     */
     public static $validationMessages = array();
 
-    # mensagens das validações
+    /**
+     * Validation set os messages when fails
+     * @var array
+     */
     public static $messages = array();
+
+    /**
+     * Setting if the model is 'sluggable'
+     * When setting this to true, the save and update methos are going to make and slug before save
+     */
+    protected $sluggable = false;
+
+    /**
+     * Sluggable attribute
+     * @var  string
+     */
+    protected $sluggableAttr = 'title';
 
     /**
      * Método de valicação
      */
-    public static function validate( $fields )
+    public static function validate($fields)
     {
         $validation = Validator::make( $fields, static::$validationRules, static::$validationMessages );
 
-        if( $validation->fails() )
-        {
+        if ($validation->fails()) {
             static::$messages = $validation->messages()->getMessages();
             
             return false;
@@ -37,8 +58,7 @@ class Eloquent extends Model
 
     public function setDataFromArray(array $data)
     {
-        foreach( $data as $attribute => $value )
-        {
+        foreach ($data as $attribute => $value) {
             $this->$attribute = $value;
         }
     }
@@ -50,14 +70,16 @@ class Eloquent extends Model
      * @param   $id    int [com o ID atual]
      * @return  string [HTML Select]
      */
-    public static function HTMLOptions( $id = null, $field = 'title', $all = null )
+    public static function HTMLOptions($id = null, $field = 'title', $all = null)
     {
-        $all  = is_null( $all ) ? static::all() : $all;
-        $html = '';
+        $class      = get_called_class();
+        $object     = new $class;
+        $primaryKey = $object->primaryKey;
+        $all        = is_null($all) ? static::all() : $all;
+        $html       = '';
 
-        foreach( $all as $item )
-        {
-            $html .= '<option value="'.$item->id.'" '.( $id == $item->id ? 'selected="selected"' : null ).'>'.$item->$field.'</option>';
+        foreach ($all as $item) {
+            $html .= '<option value="'.$item->$primaryKey.'" '.( $id == $item->$primaryKey ? 'selected="selected"' : null ).'>'.$item->$field.'</option>';
         }
 
         return $html;
@@ -70,17 +92,25 @@ class Eloquent extends Model
     {
         Event::fire(get_called_class() . '.change', array('update'));
 
+        if ($this->isSluggable()) {
+            $this->makeSlug();
+        }
+
         return parent::update($attributes);
     }
 
     /**
      * Reescrevendo a função Model::save para disparar um evento ao deletar
      */
-    public function save()
+    public function save(array $options = array())
     {
         Event::fire(get_called_class() . '.change', array('save'));
 
-        return parent::save();
+        if ($this->isSluggable()) {
+            $this->makeSlug();
+        }
+
+        return parent::save($options);
     }
 
     /**
@@ -99,19 +129,40 @@ class Eloquent extends Model
      * @access  public 
      * @return  void
      */
-    public function setFields( array $fields = array() )
+    public function setFields(array $fields = array())
     {
-        if( !is_array($this->fields) )
-        {
+        if (! is_array($this->fillable) && count($this->fillable) == 0) {
             return;
         }
         
-        foreach( $this->fields as $field )
-        {
-            if( isset( $fields[ $field ] ) )
-            {
+        foreach ($this->fillable as $field) {
+            if (isset($fields[$field])) {
                 $this->$field = $fields[ $field ];
             }
+        }
+    }
+
+    /**
+     * Returning if this model is sluggable
+     */
+    protected function isSluggable()
+    {
+        return $this->sluggable;
+    }
+
+    /**
+     * Criando o slug do produto
+     */
+    public function makeSlug()
+    {
+        $attr       = $this->sluggableAttr;
+        $slugify    = new Slugify(Slugify::MODEARRAY);
+        $this->slug = $slugify->slugify($this->$attr);
+        $count      = 0;
+
+        while ($this->where('slug', '=', $this->slug)->where('id', '<>', $this->id)->first()) {
+            $count++;
+            $this->slug = $slugify->slugify($this->$attr) . '-' . $count;
         }
     }
 }